mudanças nas classes do pacote financas e no FileHandler

+++ KappaFinancesDesktop/KappaFinances/src/Financas/Financa.java
@@ -52,7 +52,7 @@ public abstract class Financa
         return this.data;
     }
     
-    public void setValue(int value)
+    public void setValue(double value)
     {
         this.value = value;
     }
